Code restructured and refactored
Update readme
