+++ utils/utils.go
@@ -2,6 +2,7 @@ package utils
 
 import (
 	"os"
+	"regexp"
 )
 
 func WriteToFile(b []byte, out string) error {
@@ -12,4 +13,10 @@ func WriteToFile(b []byte, out string) error {
 	defer file.Close()
 	_, err = file.Write(b)
 	return err
+}
+
+func GetTokensFromStr(str string) []string {
+	re := regexp.MustCompile(`[\p{L}\d]+`) // find also unicode characters
+	tokens := re.FindAllString(str, -1)
+	return tokens
 }
